Add files via upload
fix_slab_poscar.sh: Is a bash script to fix the atoms of a slab. When the surface of the relaxed structure of the slab is modified, it is useful to fix rest of the atoms and keep only surface flexible. This script fix atoms in the range of lower_limit to upper_limit and rest of the atoms are kept flexible. 
./fix_slab_poscar.sh POSCAR lower_limit upper_limit 
new file will be generated fix_POSCAR, that will have information of fixed and flexible atoms.